requirements.txt
autopep8==1.7.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
docopt==0.6.2
et-xmlfile==1.1.0
fillpdf==0.7.2
Flask==2.2.2
Flask-Excel==0.0.7
fpdf==1.7.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lml==0.1.0
MarkupSafe==2.1.1
numpy==1.23.3
openpyxl==3.0.10
pandas==1.5.0
pdf2image==1.16.0
pdfrw2==0.5.0
Pillow==9.2.0
pipreqs==0.4.11
pycodestyle==2.9.1
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
PyMuPDF==1.20.2
pyPdf==1.13
PyPDF2==2.10.9
PyPDF4==1.27.0
pypdfium2==3.2.0
pypdftk==0.5
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
texttable==1.6.4
toml==0.10.2
urllib3==1.26.12
Werkzeug==2.2.2
yarg==0.1.9
